Compatibility with x11rb 0.13
The new version of x11rb changed HierarchyInfo.flags from a u32 to a
HierarchyMask. HierarchyMask implements Into<u32>.

Use of u32::from on info.flags is a no-op with x11rb 0.10 and lower
because flags is already a u32. With x11rb 0.13, this will convert
the flags to a u32.